Revert "Add integration tests for id token encryption"
