+++ modules/integration/tests-integration/tests-backend/src/test/java/org/wso2/identity/integration/test/oauth2/OAuth2ServiceImplicitGrantTestCase.java
@@ -193,8 +193,7 @@ public class OAuth2ServiceImplicitGrantTestCase extends OAuth2ServiceAbstractInt
 		                        response.getFirstHeader(OAuth2Constant.HTTP_RESPONSE_HEADER_LOCATION);
 		Assert.assertNotNull(locationHeader, "Approval Location header is null.");
 
-		accessToken = DataExtractUtil.extractParamFromURIFragment(locationHeader.getValue(),
-				OAuth2Constant.ACCESS_TOKEN);
+		accessToken = DataExtractUtil.extractAccessTokenFromQueryString(locationHeader.getValue());
 		Assert.assertNotNull(accessToken, "Access token is null.");
 		EntityUtils.consume(response.getEntity());
 	}
